# Conflicts:
#	.gradle/8.11.1/executionHistory/executionHistory.lock
#	.gradle/8.11.1/fileHashes/fileHashes.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	build/reports/problems/problems-report.html
#	src/main/java/com/skrookies/dahaezlge/repository/book/BookRepository.java
#	src/main/java/com/skrookies/dahaezlge/service/book/BookService.java

+++ src/main/java/com/skrookies/dahaezlge/controller/qna/Dto/QnaDto.java
@@ -12,9 +12,10 @@ import java.time.LocalDateTime;
 @ToString
 
 public class QnaDto {
-    private Integer qna_id;
+    private Long qna_id;
     private String qna_title;
     private String qna_body;
     private String qna_user_id;
     private LocalDateTime qna_created_at;
+
 }
